fix: correciones y optimización en la gestión de materiales

+++ inventario_de_sanidad/app/Http/Controllers/LoginController.php
@@ -42,12 +42,15 @@ class LoginController extends Controller
 
     public function logout()
     {
+        if (Cookie::get('TYPE') === 'admin') {
+            StorageFacade::disk('public')->deleteDirectory('temp');
+            Cookie::queue(Cookie::forget('materialsAddBasket'));
+            Cookie::queue(Cookie::forget('materialsBasket'));
+        }
+
         Cookie::queue(Cookie::forget('USERPASS'));
         Cookie::queue(Cookie::forget('NAME'));
         Cookie::queue(Cookie::forget('TYPE'));
-        Cookie::queue(Cookie::forget('materialsAddBasket'));
-        Cookie::queue(Cookie::forget('materialsBasket'));
-        StorageFacade::disk('public')->deleteDirectory('temp');
         return redirect()->route('login.form');
     }
 }
